Rename berconpy/events.py to parser.py
